requirements.txt
anyio (2.1.0)
argon2-cffi (20.1.0)
async-generator (1.10)
attrs (20.3.0)
Babel (2.9.0)
backcall (0.2.0)
beautifulsoup4 (4.9.3)
bleach (3.3.0)
certifi (2020.12.5)
cffi (1.14.5)
chardet (4.0.0)
contextvars (2.4)
cycler (0.10.0)
dataclasses (0.8)
decorator (4.4.2)
defusedxml (0.6.0)
demjson (2.2.4)
entrypoints (0.3)
idna (2.10)
imageio (2.9.0)
immutables (0.15)
importlib-metadata (3.4.0)
ipykernel (5.4.3)
ipython (7.16.1)
ipython-genutils (0.2.0)
jedi (0.18.0)
Jinja2 (2.11.3)
json5 (0.9.5)
jsonschema (3.2.0)
jtutils (0.0.8)
jupyter-client (6.1.11)
jupyter-core (4.7.1)
jupyter-server (1.3.0)
jupyterlab (3.0.7)
jupyterlab-pygments (0.1.2)
jupyterlab-server (2.2.1)
kiwisolver (1.3.1)
leven (1.0.4)
MarkupSafe (1.1.1)
matplotlib (3.3.4)
mistune (0.8.4)
nbclassic (0.2.6)
nbclient (0.5.2)
nbconvert (6.0.7)
nbformat (5.1.2)
nest-asyncio (1.5.1)
networkx (2.5)
nose (1.3.7)
notebook (6.2.0)
numpy (1.19.5)
packaging (20.9)
pandas (1.1.5)
pandocfilters (1.4.3)
parso (0.8.1)
pexpect (4.8.0)
pickleshare (0.7.5)
Pillow (8.1.0)
pip (9.0.1)
pkg-resources (0.0.0)
prometheus-client (0.9.0)
prompt-toolkit (3.0.16)
ptyprocess (0.7.0)
pycparser (2.20)
Pygments (2.8.0)
pyparsing (2.4.7)
pyrsistent (0.17.3)
python-csv (0.0.13)
python-dateutil (2.8.1)
pytz (2021.1)
PyWavelets (1.1.1)
pyzmq (22.0.3)
requests (2.25.1)
scikit-image (0.17.2)
scipy (1.5.4)
selenium (3.141.0)
Send2Trash (1.5.0)
setuptools (39.0.1)
six (1.15.0)
sniffio (1.2.0)
soupsieve (2.2)
terminado (0.9.2)
testpath (0.4.4)
tifffile (2020.9.3)
torch (1.7.1)
torchvision (0.8.2)
tornado (6.1)
traitlets (4.3.3)
typing-extensions (3.7.4.3)
urllib3 (1.26.3)
wcwidth (0.2.5)
webencodings (0.5.1)
xlrd (2.0.1)
xmltodict (0.12.0)
zipp (3.4.0)
